4_playwright/requirement.txt
pytest-playwright==0.4.3